Modified Survival analysis and RFM analysis functionality
