+++ app_components/ds/outputs/CoxPH_summary.csv
@@ -1,0 +1,4 @@
+coef,exp(coef),se(coef),coef lower 95%,coef upper 95%,exp(coef) lower 95%,exp(coef) upper 95%,cmp to,z,p,-log2(p)
+0.0004609485118021054,1.0004610547648924,0.00266550809977634,-0.004763351364259318,0.005685248387863528,0.9952479754022575,1.0057014400825386,0.0,0.1729308238983716,0.8627058026819421,0.21305943510502265
+0.09076521111517402,1.095011878456575,0.09555198164513436,-0.09651323156072161,0.27804365379106966,0.9079978832911382,1.3205438426958716,0.0,0.9499040161434048,0.34216102597662906,1.5472526566042715
+-0.00021257343671208148,0.9997874491554201,0.00032065070265292837,-0.0008410372655292831,0.0004158903921051201,0.9991593163071821,1.0004159768865046,0.0,-0.6629439291831851,0.5073664655999902,0.9788999273014225
