+++ app_components/ds/outputs/Exponential_summary.csv
@@ -1,0 +1,2 @@
+coef,se(coef),coef lower 95%,coef upper 95%,cmp to,z,p,-log2(p)
+797.3385750160919,37.882697726199936,723.0898518355226,871.5872981966612,0.0,21.04756585127376,2.4072956369753885e-98,324.28153997068983
