+++ app_components/ds/outputs/LogLogistic_summary.csv
@@ -1,0 +1,3 @@
+coef,se(coef),coef lower 95%,coef upper 95%,cmp to,z,p,-log2(p)
+713.3088099408075,12.235826138943533,689.3270313873843,737.2905884942306,1.0,58.21501563132783,0.0,inf
+4.321966415827676,0.181956750880509,3.9653377373579515,4.678595094297401,1.0,18.256901157842787,1.823512228877332e-74,244.95595914692552
