+++ app_components/ds/outputs/Weibull_summary.csv
@@ -1,0 +1,3 @@
+coef,se(coef),coef lower 95%,coef upper 95%,cmp to,z,p,-log2(p)
+806.5317657479931,13.694990072170423,779.6900784379054,833.3734530580807,1.0,58.819448681814926,0.0,inf
+2.845209659274725,0.10082440416467507,2.6475974583492516,3.0428218602001986,1.0,18.301220568196666,8.091563945854963e-75,246.12818854135483
