+++ app_components/ds/requirements.txt
@@ -1,8 +1,7 @@
-pandas
-numpy
-faker
-seaborn
-matplotlib
-faker
-lifelines
-google-generativeai
+pandas==2.2.2
+numpy==1.26.4
+faker==25.2.0
+seaborn==0.13.2
+matplotlib==3.8.4
+lifelines==0.28.0
+google-generativeai==0.4.1
